Reorganize random functions
